Added more statistics

+++ AIAD/src/stats/StatisticsGeneral.java
@@ -1,8 +1,8 @@
 package stats;
 
 public class StatisticsGeneral extends StatisticsInfo {
-    private int maxWait;
-    private int minWait;
+    private long maxWait;
+    private long minWait;
 
     public StatisticsGeneral(int maxWait, int minWait) {
         super(StatisticsType.GENERAL);
@@ -10,19 +10,19 @@ public class StatisticsGeneral extends StatisticsInfo {
         this.minWait = minWait;
     }
 
-    public int getMaxWait() {
+    public long getMaxWait() {
         return maxWait;
     }
 
-    public void setMaxWait(int maxWait) {
+    public void setMaxWait(long maxWait) {
         this.maxWait = maxWait;
     }
 
-    public int getMinWait() {
+    public long getMinWait() {
         return minWait;
     }
 
-    public void setMinWait(int minWait) {
+    public void setMinWait(long minWait) {
         this.minWait = minWait;
     }
 }
